% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_warning_data.R
\docType{data}
\name{box_warning_data}
\alias{box_warning_data}
\title{box_warning_data}
\format{The data file is a list of 1392 elements and each element is a character string.}
\usage{
box_warning_data
}
\description{
box_warning_data
}
\keyword{box}
\keyword{data}
\keyword{warning}
