% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxWarnWC.R
\name{boxWarnWC}
\alias{boxWarnWC}
\title{Create a word cloud for the box warning section}
\usage{
boxWarnWC(box_warning, language)
}
\arguments{
\item{box_warning}{contains all boxed warnings of all drugs}

\item{language}{contains key strings}
}
\description{
Generate a word cloud with the most frequently appeared words
             in the box warning section
}
\examples{
data(box_warning_data)
data(keystrings)
boxWarnWC(box_warning_data, keystrings)
}
\keyword{box}
\keyword{cloud}
\keyword{warnings}
\keyword{word}
