% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contraindications_data.R
\docType{data}
\name{contraindications_data}
\alias{contraindications_data}
\title{contraindications_data}
\format{The data file is a list of 3386 elements and each element is a character string.}
\usage{
contraindications_data
}
\description{
contraindications_data
}
\keyword{contraindications}
\keyword{data}
