% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contraWC.R
\name{contraWC}
\alias{contraWC}
\title{Create a word cloud for the contraindications section}
\usage{
contraWC(contra, language)
}
\arguments{
\item{contra}{contains all contraindications of drugs}

\item{language}{contains key strings}
}
\description{
Generate a word cloud with the most frequently appeared words
             in the contraindications section
}
\examples{
data(contraindications_data)
data(keystrings)
contraWC(contraindications_data, keystrings)
}
\keyword{cloud}
\keyword{contraindications}
\keyword{word}
