% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugWarningVA.R
\docType{package}
\name{DrugWarningVA}
\alias{DrugWarningVA}
\alias{DrugWarningVA-package}
\title{A drug warning visualization and analysis package using R}
\description{
A collection of functions for illustration. See examples
for functions \code{\link{boxWarnWC}} and \code{\link{contraWC}} 
and \code{\link{ingredientTable}}
and \code{\link{langModelTable}} and \code{\link{languageModel}}
and \code{\link{normalizeFrequency}} and \code{\link{ourStopWords}} and \code{\link{processCorpus}}
and \code{\link{visualizeLangModel}} and \code{\link{warningWC}} in addition to datasets \code{box_warning_data}
\code{contraindications_data}
\code{generic_name}
\code{keystrings}
\code{raw_box_warning_data}
\code{warning_data}
}
\references{
Julia Silge and David Robinson's book at \url{https://www.tidytextmining.com/}
}
\seealso{
\code{\link{boxWarnWC}}

\code{\link{contraWC}}

\code{\link{ingredientTable}}

\code{\link{langModelTable}}

\code{\link{languageModel}}

\code{\link{normalizeFrequency}}

\code{\link{ourStopWords}}

\code{\link{processCorpus}}

\code{\link{visualizeLangModel}}

\code{\link{warningWC}}
}
