% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_name.R
\docType{data}
\name{generic_name}
\alias{generic_name}
\title{generic_name}
\format{The data file is a list of 20000 elements and each element is a character string.}
\usage{
generic_name
}
\description{
generic_name
}
\keyword{data}
\keyword{generic}
\keyword{name}
