% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingredientTable.R
\name{ingredientTable}
\alias{ingredientTable}
\title{Genereate the frequency table of different ingredients, which count the number of time each ingredient
contained in box warning for a drug, which can be seen as an indicator of danger.}
\usage{
ingredientTable(generic_name, box_warning)
}
\arguments{
\item{generic_name}{contains all ingredients of drugs}

\item{box_warning}{contains all boxed warnings of drugs}
}
\value{
a table
}
\description{
Genereate the frequency table of different ingredients, which count the number of time each ingredient
             contained in box warning for a drug, which can be seen as an indicator of danger.
}
\examples{
data(generic_name)
data(raw_box_warning_data)
ingredientTable(generic_name, raw_box_warning_data)
}
\keyword{danger}
\keyword{each}
\keyword{for}
\keyword{ingradient}
