% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keystrings.R
\docType{data}
\name{keystrings}
\alias{keystrings}
\title{keystrings}
\format{it's Factor with 362 levels indicate the key string which we are looking for}
\usage{
keystrings
}
\description{
keystrings
}
\keyword{data}
\keyword{keystrings}
