% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langModelTable.R
\name{langModelTable}
\alias{langModelTable}
\title{Create a language model for the three warning sections}
\usage{
langModelTable(box_warning, contra, warning, language)
}
\arguments{
\item{box_warning}{contains all boxed warnings of drugs}

\item{contra}{contains all contraindications of drugs}

\item{warning}{contains all warnings of drugs}

\item{language}{contains key strings}
}
\value{
a table
}
\description{
Classify every key string as severe, moderate, or mild
             and visualize that result
}
\examples{
data(box_warning_data)
data(contraindications_data)
data(warning_data)
data(keystrings)
langModelTable(box_warning_data, contraindications_data, warning_data, keystrings)
}
\keyword{language}
\keyword{model}
\keyword{table}
