% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languageModel.R
\name{languageModel}
\alias{languageModel}
\title{Create a language model for the three warning sections}
\usage{
languageModel(box_warning, contraindications, warnings, language, word)
}
\arguments{
\item{box_warning}{contains all boxed warnings of drugs}

\item{contraindications}{contains all contraindications of drugs}

\item{warnings}{contains all warnings of drugs}

\item{language}{contains key strings}

\item{word}{a word string}
}
\value{
a row vector
}
\description{
Classify every key string as severe, moderate, or mild
             and visualize that result
}
\examples{
data(box_warning_data)
data(contraindications_data)
data(warning_data)
data(keystrings)
languageModel(box_warning_data, contraindications_data, warning_data, keystrings, "death")
}
\keyword{classification}
\keyword{language}
\keyword{model}
\keyword{warnings}
