% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeFrequency.R
\name{normalizeFrequency}
\alias{normalizeFrequency}
\title{Normalize the frequency in a given data set}
\usage{
normalizeFrequency(data)
}
\arguments{
\item{data}{counts the number of occurrences of keystrings}
}
\value{
a data drame
}
\description{
Normalize the frequency in a given data set
}
\examples{
data(box_warning_data)
data(keystrings)
bw_keystring_count <- processCorpus(box_warning_data, keystrings)
normalizeFrequency(bw_keystring_count)
}
\keyword{frequency}
\keyword{normalize}
