% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ourStopWords.R
\name{ourStopWords}
\alias{ourStopWords}
\title{Create a vector of stop words for generating word cloud}
\usage{
ourStopWords()
}
\description{
Create a vector of stop words which are to be removed
             from the corpus
}
\examples{
ourStopWords()
}
\keyword{stop}
\keyword{words}
