% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCorpus.R
\name{processCorpus}
\alias{processCorpus}
\title{Create and process a corpus}
\usage{
processCorpus(data, language)
}
\arguments{
\item{data}{a list of items to be converted to a corpus item}

\item{language}{contains key strings}
}
\value{
a matrix that contains the number of occurrences 
        in key strings
}
\description{
Generate a corpus, remove punctuations/ numbers/ stopwords,
             turn all words to lower cases and strip any white spaces
}
\examples{
data(box_warning_data)
data(keystrings)
processCorpus(box_warning_data, keystrings)
}
\keyword{corpus}
