% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_box_warning_data.R
\docType{data}
\name{raw_box_warning_data}
\alias{raw_box_warning_data}
\title{raw_box_warning_data}
\format{it's long list with 20000 elements}
\usage{
raw_box_warning_data
}
\description{
raw_box_warning_data
}
\keyword{data}
