% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeLangModel.R
\name{visualizeLangModel}
\alias{visualizeLangModel}
\title{Create a language model for the three warning sections}
\usage{
visualizeLangModel(box_warning, contraindications, warnings, language)
}
\arguments{
\item{box_warning}{contains all boxed warnings of drugs}

\item{contraindications}{contains all contraindications of drugs}

\item{warnings}{contains all warnings of drugs}

\item{language}{contains key strings}
}
\description{
Classify every key string as severe, moderate, or mild
             and visualize that result
}
\examples{
data(box_warning_data)
data(contraindications_data)
data(warning_data)
data(keystrings)
visualizeLangModel(box_warning_data, contraindications_data, warning_data, keystrings)
}
\keyword{language}
\keyword{model}
\keyword{visualization}
