% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warningWC.R
\name{warningWC}
\alias{warningWC}
\title{Create a word cloud for the warning sections}
\usage{
warningWC(warning, language)
}
\arguments{
\item{warning}{contains all warnings of drugs}

\item{language}{contains key strings}
}
\description{
Generate a word cloud with the most frequently appeared words
             in the warning sections
}
\examples{
data(warning_data)
data(keystrings)
boxWarnWC(warning_data, keystrings)
}
\keyword{cloud}
\keyword{precautions}
\keyword{warnings}
\keyword{word}
